Fixed the Snipr link...
The &amp; was not being encoded properly and was directing the cURL command to pull the information from an entirely different page which was not just a blank page with the short URL. So I removed &amp; and replaced with & instead. Now it directs to blank page with short URL so that the command doesn't insert an entire other website inside the user's website.