Added check to the curl check for safe_mode enabled and open_basedir settings that cause redirects to fail. When either is set, a warning is included in Detecting Media Information.
Better error message for when SSL is not supported in the cURL library if enabled.

Better memory_limit logic, if the value is set to zero the defaults are used based on the version of PHP installed. -1 is also recongized as having no memory limit for scripts.

Safe mode and open_basedir settings are now included in the email for better reporting.