Update help file with notes about permissions for the cache directory
Delete the Zip file which was committed by accident

Update to readme file for product description change

Update readme for the v1.2.1 release

Update product description

Add a ‘cache’ directory with a stub file

This is so that we can easily package in a cache directory and not
require users to create it in the event that the plugin does not have
permissions to do so.

Try to create cache directory if it does not exist

If the cache directory does not exist and does not have the proper
ppermissions, then users get an error to that effect.  Here we test
for the existence of the cache directory and then try to create it if
needed.

This itself can fail, if either the plugin script is lacking proper
permissions, of if the directory containing the cache directory does
not have the right permissions.  To help with this, these changes also
try to be a little more aware of possible permission failings,
checking the results of file writes and such.