Version 1.0.3.
Compacted some code, and added some brackets in places that needed them.
Changed the installation function to only insert sample data if there is not already data in the database.
Fixed some problems with double-commas.
Replaced all instances of the deprecated ereg_replace function to preg_replace.
Modified some of the core processing code to make it a little bit faster.